👌 IMPROVE: Improve timetable layout
