+++ src/components/Timetable/TimetableCourseItem.jsx
@@ -6,13 +6,16 @@ import { tracks, rows, slots } from './timetableData'
 import TimetableLectureItem from './TimetableLectureItem'
 
 const TimetableCourseItem = ({ id, courseCode }) => {
+  console.log(courseCode)
   let courseSlots = useSelector(selectCourseSlotsByCourseCode(courseCode))
+  if (!courseSlots) return null
   courseSlots = courseSlots.map((slot) => slots[slot])
 
   return (
     <>
       {courseSlots.map(({ row, track }) => (
         <TimetableLectureItem
+          key={track}
           id={id}
           title={courseCode}
           track={tracks[track - 1]}
